+++ src/components/Calculator.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import PropTypes from 'prop-types';
 import '../styles/Calculator.css';
 import calculate from '../logic/calculate';
+import Button from './Button';
 
 const Calculator = () => {
   const [displayValue, setDisplayValue] = useState('0');
@@ -56,31 +56,6 @@ const Calculator = () => {
       </div>
     </div>
   );
-};
-
-const Button = ({
-  label, color, doubleSize, onClick,
-}) => {
-  const buttonStyle = {
-    backgroundColor: color === 'orange' ? '#FF8C00' : '#E0E0E0',
-  };
-
-  return (
-    <button className={`button ${doubleSize ? 'zero' : ''}`} style={buttonStyle} type="button" onClick={() => onClick(label)}>
-      {label}
-    </button>
-  );
-};
-
-Button.propTypes = {
-  label: PropTypes.string.isRequired,
-  color: PropTypes.string.isRequired,
-  doubleSize: PropTypes.bool,
-  onClick: PropTypes.func.isRequired,
-};
-
-Button.defaultProps = {
-  doubleSize: false,
 };
 
 export default Calculator;
